+++ src/primitives/block.h
@@ -35,7 +35,7 @@ public:
     uint32_t nTime;
     uint32_t nBits;
     uint32_t nNonce;
-    uint32_t nAccumulatorChecksum;
+    uint256 nAccumulatorChecksum;
 
     CBlockHeader()
     {
